Add new unit test. Modify import of property not to throw exception if the property mapper is not loaded
git-svn-id: http://forge2.pole-aquinetic.fr/svnroot/fesapi/trunk@273 553bd026-d4ad-425d-a02a-e125c7656c9a

+++ test/AbstractTest.cpp
@@ -40,7 +40,7 @@ void AbstractTest::deserialize() {
 	vector<string> warningSet = this->epcDoc->getWarnings();
 	if (warningSet.size() > 0) {
 		cout << "EPC document " << this->epcDoc->getName() << ".epc deserialized with " << warningSet.size() << " warning(s)" << endl;
-		for (unsigned int i=0; i<warningSet.size(); ++i){
+		for (size_t i=0; i<warningSet.size(); ++i){
 			cout << "Warning " << i+1 << ": " << warningSet[i] << endl;
 		}
 		cout << endl;
